// src/multiplayer/generated/ShipState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ShipState = { id: string, tick: number, x: number, y: number, z: number, 
/**
 * Radians about the bow axis; positive lifts the starboard side (a heel to port).
 */
roll: number, 
/**
 * Radians; positive lifts the bow.
 */
pitch: number, 
/**
 * Radians, clockwise seen from above: 0 steams toward -z, pi/2 toward +x.
 */
heading: number, 
/**
 * Physical metres per second along the bow, negative astern. World motion is
 * `mobility::SHIP_PACE` times this.
 */
speed: number, swaySpeed: number, rudder: number, yawRate: number, distance: number, verticalSpeed: number, waveHeave: number, driftX: number, driftZ: number, };